Enhances homepage with scroll animations
Implements scroll animations to improve user experience on the homepage.
Wraps sections of the homepage within the new ScrollAnimation component to
animate elements into view as the user scrolls.
Also fixes a bug in the resume download analytics tracking, and a bug on SocialShare that prevents sharing if navigator is not defined